LPS-X Fix circular dependencies "SearchEngineHelperImpl -> tracker of SearchEngineConfigurator -> ElasticsearchEngineConfigurator -> SearchEngineHelperImpl"
