+++ modules/apps/portal-search/portal-search/src/main/java/com/liferay/portal/search/internal/SearchEngineHelperImpl.java
@@ -24,7 +24,6 @@ import com.liferay.portal.kernel.search.Document;
 import com.liferay.portal.kernel.search.Indexer;
 import com.liferay.portal.kernel.search.IndexerRegistryUtil;
 import com.liferay.portal.kernel.search.SearchEngine;
-import com.liferay.portal.kernel.search.SearchEngineConfigurator;
 import com.liferay.portal.kernel.search.SearchEngineHelper;
 import com.liferay.portal.kernel.search.queue.QueuingSearchEngine;
 import com.liferay.portal.kernel.util.ClassUtil;
@@ -41,13 +40,9 @@ import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
 
 import org.osgi.framework.BundleContext;
-import org.osgi.framework.ServiceReference;
 import org.osgi.service.component.annotations.Activate;
 import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Deactivate;
 import org.osgi.service.component.annotations.Modified;
-import org.osgi.util.tracker.ServiceTracker;
-import org.osgi.util.tracker.ServiceTrackerCustomizer;
 
 /**
  * @author Michael C. Han
@@ -301,49 +296,6 @@ public class SearchEngineHelperImpl implements SearchEngineHelper {
 		Collections.addAll(
 			_excludedEntryClassNames,
 			searchEngineHelperConfiguration.excludedEntryClassNames());
-
-		_serviceTracker = new ServiceTracker<>(
-			bundleContext, SearchEngineConfigurator.class,
-			new ServiceTrackerCustomizer
-				<SearchEngineConfigurator, SearchEngineConfigurator>() {
-
-				@Override
-				public SearchEngineConfigurator addingService(
-					ServiceReference<SearchEngineConfigurator>
-						serviceReference) {
-
-					SearchEngineConfigurator searchEngineConfigurator =
-						bundleContext.getService(serviceReference);
-
-					searchEngineConfigurator.afterPropertiesSet();
-
-					return searchEngineConfigurator;
-				}
-
-				@Override
-				public void modifiedService(
-					ServiceReference<SearchEngineConfigurator> serviceReference,
-					SearchEngineConfigurator searchEngineConfigurator) {
-				}
-
-				@Override
-				public void removedService(
-					ServiceReference<SearchEngineConfigurator> serviceReference,
-					SearchEngineConfigurator searchEngineConfigurator) {
-
-					searchEngineConfigurator.destroy();
-
-					bundleContext.ungetService(serviceReference);
-				}
-
-			});
-
-		_serviceTracker.open();
-	}
-
-	@Deactivate
-	protected void deactivate() {
-		_serviceTracker.close();
 	}
 
 	private static final Log _log = LogFactoryUtil.getLog(
@@ -357,7 +309,5 @@ public class SearchEngineHelperImpl implements SearchEngineHelper {
 		new HashMap<>();
 	private final Map<String, SearchEngine> _searchEngines =
 		new ConcurrentHashMap<>();
-	private volatile ServiceTracker
-		<SearchEngineConfigurator, SearchEngineConfigurator> _serviceTracker;
 
 }
